refactor Manage Stone Code Activity
