+++ app/build.gradle
@@ -1,5 +1,8 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
+plugins {
+    id 'com.android.application'
+    id 'org.jetbrains.kotlin.android'
+    alias(libs.plugins.compose.compiler)
+}
 apply from: '../gertec/gertec-signing-config.gradle'
 apply from: '../positivo/positivo-signing-config.gradle'
 
@@ -23,6 +26,9 @@ android {
         versionName "1.0"
     }
 
+    buildFeatures {
+        compose true
+    }
     buildTypes {
         release {
             minifyEnabled false
@@ -51,6 +57,23 @@ android {
 }
 
 dependencies {
+
+    def composeBom = platform('androidx.compose:compose-bom:2024.06.00')
+    implementation composeBom
+    implementation 'androidx.compose.material3:material3'
+    implementation 'androidx.compose.foundation:foundation'
+    implementation 'androidx.compose.ui:ui'
+    implementation 'androidx.compose.ui:ui-tooling-preview'
+    implementation 'androidx.activity:activity-compose:1.9.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.8.3'
+
+    implementation("io.insert-koin:koin-core:3.1.6")
+    implementation("io.insert-koin:koin-android:3.1.6")
+    implementation("io.insert-koin:koin-androidx-compose:3.1.6")
+
+    debugImplementation 'androidx.compose.ui:ui-tooling'
+    androidTestImplementation composeBom
+
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.7.0'
     implementation 'com.github.permissions-dispatcher:permissionsdispatcher:4.8.0'
@@ -69,8 +92,6 @@ dependencies {
     implementation "br.com.stone:stone-sdk-posandroid-sunmi:$stone_sdk_version"
     implementation "androidx.activity:activity-ktx:1.9.1"
     implementation "androidx.fragment:fragment-ktx:1.8.2"
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     testImplementation "io.mockk:mockk:1.13.12"
     testImplementation "org.junit.platform:junit-platform-launcher:1.10.0"
