+++ app/src/main/java/br/com/stonesdk/sdkdemo/activities/manageStoneCode/ManageStoneCodeActivity.kt
@@ -1,26 +1,23 @@
-package br.com.stonesdk.sdkdemo.activities
+package br.com.stonesdk.sdkdemo.activities.manageStoneCode
 
 import android.os.Bundle
 import android.util.Log
-import android.view.View
 import android.widget.AdapterView.OnItemClickListener
 import android.widget.ArrayAdapter
 import android.widget.EditText
 import android.widget.ListView
 import android.widget.Toast
+import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.material3.MaterialTheme
 import br.com.stonesdk.sdkdemo.R
-import br.com.stonesdk.sdkdemo.databinding.ActivityMainBinding
+import br.com.stonesdk.sdkdemo.activities.FeatureFlag
 import br.com.stonesdk.sdkdemo.databinding.ActivityManageStoneCodeBinding
 import stone.application.interfaces.StoneCallbackInterface
 import stone.providers.ActiveApplicationProvider
 import stone.user.UserModel
 import stone.utils.Stone
 
-/**
- * @author tiago.barbosa
- * @since 10/04/2018
- */
 class ManageStoneCodeActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityManageStoneCodeBinding
@@ -30,7 +27,18 @@ class ManageStoneCodeActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        if (FeatureFlag.refactorEnabled) {
+            setContent {
+                MaterialTheme {
+                    ManageStoneCodeScreen()
+                }
+            }
+        } else {
+            onCreateConfig()
+        }
+    }
 
+    private fun onCreateConfig() {
         binding = ActivityManageStoneCodeBinding.inflate(layoutInflater)
         setContentView(binding.root)
