+++ app/src/main/res/layout/activity_manage_stone_code.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="br.com.stonesdk.sdkdemo.activities.ManageStoneCodeActivity">
+    tools:context="br.com.stonesdk.sdkdemo.activities.manageStoneCode.ManageStoneCodeActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/input_layout_password"
