+++ build.gradle
@@ -1,27 +1,32 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext {
-        kotlin_version = '2.0.0'
-        stone_sdk_version = '4.9.4'
-
-        localProp = new Properties()
-        fileName = 'local.properties'
-        if (project.rootProject.file(fileName).exists()) {
-            localProp.load(new FileInputStream(rootProject.file(fileName)))
-        }
-        packageCloudReadToken = System.env.PACKAGECLOUD_READ_TOKEN != null ? System.env.PACKAGECLOUD_READ_TOKEN  : localProp["packageCloudReadToken"] ?: ""
-        packageCloudReadTokenInternal = System.env.PACKAGECLOUD_READ_TOKEN_INTERNAL ?: localProp["packageCloudReadTokenInternal"]
-    }
-
     repositories {
-        mavenCentral()
+        gradlePluginPortal()
         google()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:8.5.2'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath libs.gradlePlugin
+        classpath libs.kotlinGradlePlugin
     }
+}
+
+plugins {
+    alias(libs.plugins.compose.compiler) apply false
+}
+
+ext {
+    kotlin_version = '2.0.0'
+    stone_sdk_version = '4.9.4'
+
+    localProp = new Properties()
+    fileName = 'local.properties'
+    if (project.rootProject.file(fileName).exists()) {
+        localProp.load(new FileInputStream(rootProject.file(fileName)))
+    }
+    packageCloudReadToken = System.env.PACKAGECLOUD_READ_TOKEN != null ? System.env.PACKAGECLOUD_READ_TOKEN : localProp["packageCloudReadToken"] ?: ""
+    packageCloudReadTokenInternal = System.env.PACKAGECLOUD_READ_TOKEN_INTERNAL ?: localProp["packageCloudReadTokenInternal"]
 }
 
 allprojects {
@@ -32,4 +37,11 @@ allprojects {
         maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
         google()
     }
+   /* configurations.configureEach {
+        resolutionStrategy.eachDependency { DependencyResolveDetails details ->
+            if (details.requested.group == "io.insert-koin") {
+                details.useVersion "3.5.6"
+            }
+        }
+    } */
 }
